% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_engine.R
\name{compute_engine}
\alias{compute_engine}
\title{Generic Compute Engine for WHO STEPS Tables}
\description{
Takes a table specification from \code{\link[=steps_table_registry]{steps_table_registry()}} and a survey
design object, and produces the survey-weighted estimates needed to fill
the standard WHO STEPS data book table.
}
\concept{compute}
