% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_engine.R
\name{.apply_denominator}
\alias{.apply_denominator}
\title{Optionally subset a design by a denominator variable}
\usage{
.apply_denominator(design, denominator, data)
}
\description{
Optionally subset a design by a denominator variable
}
\keyword{internal}
