% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{.build_cascade_table}
\alias{.build_cascade_table}
\title{Build a cascade table (also used for combined risk factors)}
\usage{
.build_cascade_table(result)
}
\description{
Build a cascade table (also used for combined risk factors)
}
\keyword{internal}
