% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{.col_label}
\alias{.col_label}
\title{Get the SPSS/haven label for a column (if any)}
\usage{
.col_label(col)
}
\arguments{
\item{col}{A vector (typically from a data frame column).}
}
\value{
Character string of the label, or "" if none.
}
\description{
Get the SPSS/haven label for a column (if any)
}
\keyword{internal}
