% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps_registry.R
\name{get_registry_by_step}
\alias{get_registry_by_step}
\title{Get table registry entries by step}
\usage{
get_registry_by_step(step)
}
\arguments{
\item{step}{STEPS step number (1, 2, or 3).}
}
\value{
A filtered list of registry entries.
}
\description{
Get table registry entries by step
}
