% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps_registry.R
\name{list_registry_sections}
\alias{list_registry_sections}
\title{List all available sections in the registry}
\usage{
list_registry_sections()
}
\value{
Character vector of unique section names.
}
\description{
List all available sections in the registry
}
