% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about.R
\name{mod_about_server}
\alias{mod_about_server}
\title{About Page Module Server}
\usage{
mod_about_server(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
About Page Module Server
}
\keyword{internal}
