% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about.R
\name{mod_about_ui}
\alias{mod_about_ui}
\title{About Page Module UI}
\usage{
mod_about_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A Shiny UI element.
}
\description{
Displays information about the stepssurvey package, authors,
and institutional affiliations.
}
\keyword{internal}
