% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clean.R
\name{mod_clean_server}
\alias{mod_clean_server}
\title{Clean Module — Server}
\usage{
mod_clean_server(id, upload_out)
}
\arguments{
\item{id}{Module namespace id.}

\item{upload_out}{Reactive list returned by \code{\link[=mod_upload_server]{mod_upload_server()}}.}
}
\value{
A \link[shiny:reactive]{shiny::reactive} returning the cleaned data frame.
}
\description{
Clean Module — Server
}
\keyword{internal}
