% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_design.R
\name{mod_design_server}
\alias{mod_design_server}
\title{Design Module — Server}
\usage{
mod_design_server(id, clean_out)
}
\arguments{
\item{id}{Module namespace id.}

\item{clean_out}{Reactive returning the cleaned data frame.}
}
\value{
A \link[shiny:reactive]{shiny::reactive} returning the survey design object.
}
\description{
Design Module — Server
}
\keyword{internal}
