% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indicators.R
\name{mod_indicators_server}
\alias{mod_indicators_server}
\title{Indicators Module — Server}
\usage{
mod_indicators_server(id, design_out)
}
\arguments{
\item{id}{Module namespace id.}

\item{design_out}{Reactive returning a survey::svydesign object.}
}
\value{
A list with two reactive elements:
\itemize{
\item \code{results}: reactive list of indicator results by domain
\item \code{key_indicators}: reactive tibble of headline key indicators
}
}
\description{
Indicators Module — Server
}
\keyword{internal}
