% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indicators.R
\name{mod_indicators_ui}
\alias{mod_indicators_ui}
\title{Indicators Module — UI}
\usage{
mod_indicators_ui(id)
}
\arguments{
\item{id}{Module namespace id.}
}
\value{
UI definition for indicators computation and display.
}
\description{
Indicators Module — UI
}
\keyword{internal}
