% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_quality.R
\name{mod_quality_ui}
\alias{mod_quality_ui}
\title{Data Quality Module — UI}
\usage{
mod_quality_ui(id)
}
\arguments{
\item{id}{Module namespace id.}
}
\description{
Data Quality Module — UI
}
\keyword{internal}
