% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{mod_upload_server}
\alias{mod_upload_server}
\title{Upload Module — Server}
\usage{
mod_upload_server(id)
}
\arguments{
\item{id}{Module namespace id.}
}
\value{
A \link[shiny:reactiveValues]{shiny::reactiveValues} with elements \code{raw}, \code{cols}, \code{config}.
}
\description{
Upload Module — Server
}
\keyword{internal}
