% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{recode_yn}
\alias{recode_yn}
\title{Recode a yes/no variable to logical}
\usage{
recode_yn(x)
}
\arguments{
\item{x}{A vector of yes/no values (numeric or character).}
}
\value{
A logical vector, with NA for unrecognized values.
}
\description{
Converts various representations of yes/no (numeric, text, case-insensitive)
to logical TRUE/FALSE values.
}
\examples{
recode_yn(c(1, 2, "yes", "no", NA))
}
\keyword{internal}
