% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{steps_colors}
\alias{steps_colors}
\title{WHO STEPS colour palette}
\usage{
steps_colors()
}
\value{
A named list of hex colour codes.
}
\description{
A named list of colours used in WHO STEPS reports and visualisations.
}
\examples{
steps_colors()$blue
}
