% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{package}
\name{stepssurvey-package}
\alias{stepssurvey}
\alias{stepssurvey-package}
\title{stepssurvey: Analyse WHO STEPS Survey Data}
\description{
A complete analysis pipeline for the WHO STEPwise Approach to NCD Risk
Factor Surveillance (STEPS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/drpakhare/stepssurvey}
  \item Report bugs at \url{https://github.com/drpakhare/stepssurvey/issues}
}

}
\author{
\strong{Maintainer}: Abhijit Pakhare \email{drpakhare@gmail.com}

Authors:
\itemize{
  \item Ankur Joshi
}

Other contributors:
\itemize{
  \item WHO STEPS R Pipeline Contributors [contributor]
}

}
