% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_builder.R
\name{table_builder}
\alias{table_builder}
\title{Generic Table Builder for WHO STEPS Data Book}
\description{
Takes computed results from \code{\link[=compute_table]{compute_table()}} or \code{\link[=compute_all_tables]{compute_all_tables()}}
and produces formatted flextable objects in the standard WHO STEPS
3-panel format (Men / Women / Both Sexes).
}
\concept{tables}
